Updated lots of things to use fully-specified integer types.  In particular, converted all column, row, plane, and channel indices and dimensions to int32.  This affects lots of files at once, which is why this is such a distressingly large single commit.  Also, many minor coding convention edits. In addition:
Manipulation.h:
Added non-const variants of operator=() to assignable views.

EdgeExtension.h:
Added LinearEdgeExtension mode.

Algorithms.h:
Added is_opaque() function.

Convolution.h:
Totally reworked ConvolutionView to better support block rasterization.

Transform.h:
Improved support for block rasterization.  Added a new mechanism for 
specifying forward and reverse mapping function types, so we can choose 
the proper method for transforming bounding boxes accordingly.

DiskImageResourceTIFF.*:
Added support for reading multi-plane images.  Take libtiff's advice 
on the strip size when writing images.  Tidied things up a bit.

DiskImageResourceHDF.*, configure.ac:
Added support for HDF-EOS files using HDF4 library.

GeoReference.*:
Added support for transverse Mercator projections.

GeoTransform.h:
Updated to match changes to Transform.

QuadTreeGenerator.h:
Reworked computation of tile file paths so it's more user-configurable.
(No change to default behavior.)  

ImageComposite.h:
Added ProgressCallback support to prepare().  Improved support for 
arbitrary block rasterization.

KMLQuadTreeGenerator.h:
Overrode default output file path behavior to treat the root node 
specially, so the root KML file has a meaningful name.  Updated 
GlobalKMLTransform to match changes to Transform.

image2kml.cc:
Improved command-line option handling.  Added support for user-specified 
projection information.  Added progress bars.  Added support for 
generating only the portion of the global quad-tree containing the image, 
to avoid unnecessary directory nesting.

+++ src/vw/FileIO/DiskImageResourcePDS.cc
@@ -211,8 +211,8 @@ void vw::DiskImageResourcePDS::read( ImageBuffer const& dest, BBox2i const& bbox
   fseek(input_file, m_image_data_offset, 0);
 
   // Grab the pixel data from the file.
-  unsigned int total_pixels = m_format.cols * m_format.rows * m_format.planes;
-  unsigned int bytes_per_pixel = 1;
+  unsigned total_pixels = (unsigned)( m_format.cols * m_format.rows * m_format.planes );
+  unsigned bytes_per_pixel = 1;
   if ( m_format.channel_type == VW_CHANNEL_UINT16 ||
        m_format.channel_type == VW_CHANNEL_INT16 ) {
     bytes_per_pixel = 2;
@@ -264,10 +264,10 @@ void vw::DiskImageResourcePDS::read( ImageBuffer const& dest, BBox2i const& bbox
       int16 valid_minimum = atoi(valid_minimum_str.c_str());
       uint8* src_row = (uint8*)src.data;
       uint8* dst_row = (uint8*)dest.data;
-      for( unsigned y=0; y<m_format.rows; ++y ) {
+      for( int32 y=0; y<m_format.rows; ++y ) {
         uint8* src_data = src_row;
         uint8* dst_data = dst_row;
-        for( unsigned x=0; x<m_format.cols; ++x ) {
+        for( int32 x=0; x<m_format.cols; ++x ) {
           if( *((int16*)src_data) < valid_minimum ) {
             memset( dst_data, 0, dst_bpp );
           }
